Move OpCandidate, OpMatcher, etc. to engine
