+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/matcher/OpMatcher.java
@@ -27,11 +27,12 @@
  * #L%
  */
 
-package org.scijava.ops.api.features;
+package org.scijava.ops.engine.matcher;
 
-import org.scijava.ops.api.OpCandidate;
 import org.scijava.ops.api.OpEnvironment;
 import org.scijava.ops.api.OpRef;
+import org.scijava.ops.api.features.MatchingConditions;
+import org.scijava.ops.engine.OpCandidate;
 
 /**
  * Finds Ops which match an {@link OpRef}.
