+++ scijava/scijava-ops-engine/src/main/java/org/scijava/ops/engine/matcher/adapt/AdaptationMatchingRoutine.java
@@ -15,8 +15,8 @@ import java.util.stream.Collectors;
 
 import org.scijava.ops.api.Hints;
 import org.scijava.ops.api.InfoChain;
-import org.scijava.ops.api.OpCandidate;
-import org.scijava.ops.api.OpCandidate.StatusCode;
+import org.scijava.ops.engine.OpCandidate;
+import org.scijava.ops.engine.OpCandidate.StatusCode;
 import org.scijava.ops.api.OpDependencyMember;
 import org.scijava.ops.api.OpEnvironment;
 import org.scijava.ops.api.OpInfo;
@@ -24,8 +24,8 @@ import org.scijava.ops.api.OpRef;
 import org.scijava.ops.api.features.BaseOpHints.Adaptation;
 import org.scijava.ops.engine.DependencyMatchingException;
 import org.scijava.ops.api.features.MatchingConditions;
-import org.scijava.ops.api.features.MatchingRoutine;
-import org.scijava.ops.api.features.OpMatcher;
+import org.scijava.ops.engine.matcher.MatchingRoutine;
+import org.scijava.ops.engine.matcher.OpMatcher;
 import org.scijava.ops.api.features.OpMatchingException;
 import org.scijava.ops.engine.matcher.impl.DefaultOpRef;
 import org.scijava.ops.engine.struct.FunctionalParameters;
